The Next Level
Fixed Timer Issues, added game over on hit, added a level 2 to start
mapping out what we need to do.
Waiting on some assets to start mapping out the next level. I need gui
for level 2, just transfer it over from the tutorial. When needing a
timer use the timerobject with the time script attached it keeps track
of time from the start of each level